pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "github-project-automation"
version = "0.1.0"
description = "Production-grade tool to create GitHub issues and manage GitHub Project (v2) fields."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name="Your Name"}]
dependencies = [
  "requests>=2.31.0",
  "python-dotenv>=1.0.1",
  "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.4.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "B", "UP"]
ignore = []
